+++ QuantizationTable.java
@@ -5,7 +5,7 @@
  */
 public class QuantizationTable extends PixelBlock {
 
-    private final int[][] DEFAULT_LUMINANCE = {
+    private final float[][] DEFAULT_LUMINANCE = {
         {16, 11, 10, 16, 24, 40, 51, 61},
         {12, 12, 14, 19, 26, 58, 60, 55},
         {14, 13, 16, 24, 40, 57, 69, 56},
@@ -16,7 +16,7 @@ public class QuantizationTable extends PixelBlock {
         {72, 92, 95, 98, 112, 100, 103, 99}
     };
 
-    private final int[][] DEFAULT_CHROMINANCE = {
+    private final float[][] DEFAULT_CHROMINANCE = {
         {17, 18, 24, 47, 99, 99, 99, 99},
         {18, 21, 26, 66, 99, 99, 99, 99},
         {24, 26, 56, 99, 99, 99, 99, 99},
@@ -34,18 +34,18 @@ public class QuantizationTable extends PixelBlock {
         channel3 = DEFAULT_CHROMINANCE;
     }
 
-    public void setConstant(int c) {
+    public void setConstant(float c) {
         for (int i = 0; i < PixelBlock.WIDTH; i++) {
             for (int j = 0; j < PixelBlock.HEIGHT; j++)
                 channel1[i][j] = c;
         }
     }
 
-    public void setLuminance(int[][] vals) {
+    public void setLuminance(float[][] vals) {
         channel1 = vals;
     }
 
-    public void setChrominance(int [][] vals) {
+    public void setChrominance(float[][] vals) {
         channel2 = vals;
         channel3 = vals;
     }
